Moved Unit into common code
