+++ game/core/src/units/units_holder.ts
@@ -26,7 +26,6 @@ import {
 } from "@heroesofcrypto/common";
 
 import { FightStateManager } from "../state/fight_state_manager";
-import { RenderableUnit } from "./renderable_unit";
 
 export class UnitsHolder {
     private readonly grid: Grid;
@@ -238,7 +237,7 @@ export class UnitsHolder {
                 if (newAmountAlive > 0) {
                     unitToDelete.increaseAmountAlive(newAmountAlive);
                     unitToDelete.decreaseAmountDied(newAmountAlive);
-                    (unitToDelete as RenderableUnit).enqueueResurrectionAnimation();
+                    unitToDelete.handleResurrectionAnimation();
                     unitToDelete.deleteAllEffects();
                     unitToDelete.deleteAllBuffs();
                     unitToDelete.deleteAllDebuffs();
